Moved login/register dialog to common components
