+++ app/admin/components/AdminApp.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 
-import AuthenticationManager from '../../main/components/AuthenticationManager'
+import AuthenticationManager from '../../common/components/AuthenticationManager'
 import Sidebar from './Sidebar'
 import StatusPanel from './StatusPanel'
 import UserPanel from './UserPanel'
