+++ app/main/actions/index.js
@@ -6,49 +6,17 @@ START actions are dispatched when the task begins, FINISH when successfully
 completed, and ERROR if any sort of error occured.
 *******************************************************************/
 
-import {createAction} from 'redux-actions'
-import HttpStatus from 'http-status'
+import { createAction } from 'redux-actions'
 import log from 'loglevel'
 
+import { createTaskThunk } from '../../common/actions'
 import * as api from './api'
 import stuffrApi from '../../stuffrapi'
 
-export {api}
-
-/* Generic thunk creator for tasks that take a significant amount of time to
-complete, e.g. those that make network requests. *Action parameters should
-be the appropriate action creators, and apiFunction is an async function
-that makes the request. apiFunction is called with the parameters given
-to the thunk and doneAction is dispatched with the apiFunction's return
-value. */
-export function createTaskThunk (apiFunction, requestAction, doneAction, errorAction) {
-  return function (...apiParams) {
-    return async function (dispatch, getState) {
-      dispatch(requestAction())
-      try {
-        const response = await apiFunction({dispatch, getState}, ...apiParams)
-        dispatch(doneAction(response))
-        return response
-      } catch (error) {
-        dispatch(errorAction(error))
-        if (error.status === HttpStatus.UNAUTHORIZED) {
-          dispatch(authorizationRequired())
-        }
-      }
-    }
-  }
-}
-
-// Places the API token where it can be retrieved for later sessions
-function storeToken (token) {
-  window.localStorage.apiToken = token
-}
+export { api }
 
 /* Event actions
-*******************/
-export const AUTHORIZATION_REQUIRED = 'AUTHORIZATION_REQUIRED'
-export const authorizationRequired = createAction(AUTHORIZATION_REQUIRED)
-
+*****************/
 export const OPEN_THING_EDITOR = 'OPEN_THING_EDITOR'
 export const openThingEditor = createAction(OPEN_THING_EDITOR)
 export const CLOSE_THING_EDITOR = 'CLOSE_THING_EDITOR'
@@ -59,66 +27,7 @@ export const CLOSE_INVENTORY_EDITOR = 'CLOSE_INVENTORY_EDITOR'
 export const closeInventoryEditor = createAction(CLOSE_INVENTORY_EDITOR)
 
 /* Task actions
-*******************/
-
-// Actions to log a user into the server
-export const LOGIN_USER__START = 'LOGIN_USER__START'
-export const loginUserStart = createAction(LOGIN_USER__START)
-export const LOGIN_USER__FINISH = 'LOGIN_USER__FINISH'
-export const loginUserFinish = createAction(LOGIN_USER__FINISH)
-export const LOGIN_USER__ERROR = 'LOGIN_USER__ERROR'
-export const loginUserError = createAction(LOGIN_USER__ERROR)
-// loginUser - Perform user login and initial client setup.
-// Parameters:
-//  email: Login ID
-//  password: Make a wild guess
-// Returns:
-//  User info object
-export const loginUser = createTaskThunk(
-  async function ({dispatch, getState}, email, password) {
-    log.info(`loginUser: Logging in user ${email}`)
-    await stuffrApi.login(email, password)
-    storeToken(stuffrApi.token)
-    return dispatch(loadUser())
-  },
-  loginUserStart, loginUserFinish, loginUserError
-)
-
-// Actions to register a new user
-export const REGISTER_USER__START = 'REGISTER_USER__START'
-export const registerUserStart = createAction(REGISTER_USER__START)
-export const REGISTER_USER__FINISH = 'REGISTER_USER__FINISH'
-export const registerUserFinish = createAction(REGISTER_USER__FINISH)
-export const REGISTER_USER__ERROR = 'REGISTER_USER__ERROR'
-export const registerUserError = createAction(REGISTER_USER__ERROR)
-// registerUser - Perform user registration
-// Parameters:
-//  email: Login ID
-//  password: Make a wild guess
-export const registerUser = createTaskThunk(
-  async function ({dispatch, getState}, newUserData) {
-    log.info('registerUser: Registering new user')
-    await stuffrApi.registerUser(newUserData)
-    storeToken(stuffrApi.token)
-    dispatch(loadUser())
-  },
-  registerUserStart, registerUserFinish, registerUserError
-)
-
-// Actions to log a user out and remove all their data from the client
-export const PURGE_USER = 'PURGE_USER'
-export const purgeUser = createAction(PURGE_USER)
-// logoutUser - Log the user out and clean up
-export const logoutUser = function () {
-  return function (dispatch) {
-    log.info('purgeUser: Purging user data')
-    stuffrApi.logout()
-    delete window.localStorage.apiToken
-    delete window.localStorage.lastInventoryId
-    // Reducers use purgeUser to clean user data from state
-    dispatch(purgeUser())
-  }
-}
+****************/
 
 // Actions to load a user and their data after login
 export const LOAD_USER__START = 'LOAD_USER__START'
